SSCS-10262 Package Restructure
