+++ src/main/java/uk/gov/hmcts/reform/sscs/model/service/hearingvalues/CaseFlags.java
@@ -1,11 +1,14 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.service.hearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Data
 @Builder(toBuilder = true)
@@ -13,10 +16,9 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class HearingLocation {
+public class CaseFlags {
 
-    private String locationId;
-    private String locationName;
-    private String locationType;
-    private String region;
+    private List<PartyFlags> flags;
+    @JsonProperty("flagAmendURL")
+    private String flagAmendUrl;
 }
