+++ src/main/java/uk/gov/hmcts/reform/sscs/model/service/hearingvalues/HearingLocation.java
@@ -1,14 +1,11 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.service.hearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @Builder(toBuilder = true)
@@ -16,9 +13,10 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class CaseFlags {
+public class HearingLocation {
 
-    private List<PartyFlags> flags;
-    @JsonProperty("flagAmendURL")
-    private String flagAmendUrl;
+    private String locationId;
+    private String locationName;
+    private String locationType;
+    private String region;
 }
