+++ src/main/java/uk/gov/hmcts/reform/sscs/model/service/hearingvalues/PanelComposition.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.service.hearingvalues;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -13,9 +13,8 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @NoArgsConstructor
-public class Navigation {
+public class PanelComposition {
 
-    private String conditionOperator;
-    private String conditionValue;
-    private String resultValue;
+    private String memberType;
+    private Integer count;
 }
