+++ src/main/java/uk/gov/hmcts/reform/sscs/model/service/hearingvalues/PartyFlagsMap.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.service.hearingvalues;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
