+++ src/test/java/uk/gov/hmcts/reform/sscs/helper/mapping/PartyFlagsMappingTest.java
@@ -11,21 +11,21 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingOptions;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.YesNo;
-import uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlags;
+import uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlags;
 
 import java.util.List;
 
 import static java.util.Objects.nonNull;
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.ADJOURN_CASE_INTERPRETER_LANGUAGE;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.DISABLED_ACCESS;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.DWP_PHME;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.DWP_UCB;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.HEARING_LOOP;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.IS_CONFIDENTIAL_CASE;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.SIGN_LANGUAGE_TYPE;
-import static uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyFlagsMap.URGENT_CASE;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.ADJOURN_CASE_INTERPRETER_LANGUAGE;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.DISABLED_ACCESS;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.DWP_PHME;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.DWP_UCB;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.HEARING_LOOP;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.IS_CONFIDENTIAL_CASE;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.SIGN_LANGUAGE_TYPE;
+import static uk.gov.hmcts.reform.sscs.model.service.hearingvalues.PartyFlagsMap.URGENT_CASE;
 
 class PartyFlagsMappingTest extends HearingsMappingBase {
