+++ src/test/java/uk/gov/hmcts/reform/sscs/model/service/ServiceHearingValuesTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -10,6 +10,7 @@ import com.networknt.schema.SpecVersion;
 import com.networknt.schema.ValidationMessage;
 import org.junit.jupiter.api.Test;
 import uk.gov.hmcts.reform.sscs.ResourceLoader;
+import uk.gov.hmcts.reform.sscs.model.service.hearingvalues.ServiceHearingValues;
 
 import java.util.Set;
